chore(MeshHTTPRoute): refactor filters
