+++ pkg/plugins/policies/meshhttproute/plugin/v1alpha1/testdata/url-rewrite.listeners.golden.yaml
@@ -37,11 +37,13 @@ resources:
                       googleRe2: {}
                       regex: .*
                     substitution: /v2
+                  timeout: 0s
               - match:
                   prefix: /v1/
                 route:
                   cluster: backend
                   prefixRewrite: /v2
+                  timeout: 0s
               - match:
                   prefix: /
                 route:
